add requests module
