+++ miamoremustachio/app/modules/helpers.js
@@ -1,4 +1,4 @@
-const { SERVER_URL, FAVICON, ERROR } = require('./constants');
+const { FAVICON, ERROR } = require('./constants');
 
 const isClientError = status => (status >= 400 && status < 500);
 const isServerError = status => (status >= 500);
@@ -8,7 +8,7 @@ module.exports.getNameFromUrl = url => url.slice(1);
 
 const { REQUEST_CLIENT_ERROR, REQUEST_SERVER_ERROR } = ERROR;
 
-const getResponseErrorMessage = function(status) {
+module.exports.getResponseErrorMessage = function(status) {
   let errorMessage;
 
   if (isClientError(status)) {
@@ -20,16 +20,4 @@ const getResponseErrorMessage = function(status) {
   }
 
   return errorMessage;
-}
-
-module.exports.requestName = function(name) {
-  return fetch(`${SERVER_URL}?name=${name}`)
-    .then(response => {
-      if (!response.ok) {
-        const errorMessage = getResponseErrorMessage(response.status);
-        throw new Error(errorMessage);
-      }
-
-      return response.json();
-    });
 }
